requirements.txt
cffi==1.15.1
cryptography==39.0.0
Pillow==9.4.0
prettytable==3.6.0
pycparser==2.21
wcwidth==0.2.6
